pyproject.toml
[project]
name = "entitled-party"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography==41.0.1",
    "idna==3.4",
    "pyjwt==2.7.0",
    "pyopenssl==23.2.0",
    "requests==2.31.0",
    "urllib3==2.0.3",
]
